[PM] Register analysis and passes to accept delete notifications.
This commit makes the PassManager register transformation passes and analysis to
receive delete notifications. Our original plan was for analysis to be
registered to accept delete notifications only when they are requested with
getAnalysis<>, but this won't work well for analysis that are used by other
analysis (and are live across the optimization pipeline).